fix: #8 reorganize tests
