+++ tests/conftest.py
@@ -1,0 +1,3 @@
+pytest_plugins = [
+  "fixtures.paprika_etl",
+]
